Mentions: Factor out the message parsing logic
